Ex4: Started exercise 4, Add score tracker, victory screen
I defined a left paddle score variable that keeps track of the number of times the ball goes off the left side of the screen. Using this variable, an if statement , a for loop and an array of objects, a block is drawn on the right side of the screen every time the ball goes of the screen. These blocks are counted as the right paddle score.
Once the number of blocks reaches to 15, the right paddle wins.