Ex5: Make a seperate method to show prey eating tracker
I made a seperate method which displays the number of preys eaten for each player. In order to literally showing the text, I embeded the code in the display method so that it is displayed just like other objects on the screen. (I did this to keep my code nice and neat).