Ex5: Add start screen, modify preys object declaration and assignment way
I added a start screen displaying an instruction and a start button so the player knows what to do to win the game. I also modified my code of side and actual preys. I replaced the previous version of my code with a for loop containing push() method and some shorter arrays to make my code cleaner and nicer.