P2: Start renaming variables + modified and added new methods
I started to rename my variables in order to have accordance with other elements of the game. Second, to make the game more interesting I am adding new features to the game. Like: if the player hit the barrier the goals disappear(I am still working on that). And in order to make the goals visible again, he should get help from the friends. The awareness method might be deleted depending on howmuch necessary it's presence would be.