Ex8: Modify target code + add second transition screen
I added if statements and ids so that by every time ball target collision, that target point is added only once to the player score.  All these settings will be reset once the target goes off the right side of screen, so that it would be counted again if the ball hits it. I also added second transition screen and third step.