P3: Add rotated screen
I added the code which after barrier ball collision will result in screen rotation. However, not all barriers do so. A random number of barriers leads to rotation and a random number of it to player health decrement. There is also a screen between two modes which warns the player about rotation.