P3: Finish third step by 90 percent
I put some of the elements of ball, targets and barriers classes in gameStructure class because I want the same variables to be used for both rotated and not rotated screen so that values are added to the same variables. Like health percent or player score. This way, the rotated and not rotated screen won't be counted as two seperate screens. After each rotation paddle is reset to it's initial value so that it doesn't touch  any barrier before the player decides to move.